backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
google-cloud-speech>=2.21.0
pyaudio>=0.2.13
numpy>=1.24.0
vertexai>=1.30.0
google-auth>=2.22.0
google-cloud-core>=2.3.0
google-generativeai>=0.3.0
openai>=1.3.0 